requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
html5lib==1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
Pillow==8.1.2
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.66
sqlparse==0.4.1
webencodings==0.5.1
xhtml2pdf==0.2.5
